create barchart component
update piechart import
