+++ src/pages/Calculator/Steps/ReviewMix/index.js
@@ -4,26 +4,18 @@
 
 import React, { useEffect, useState } from 'react';
 import Grid from '@mui/material/Grid';
-import { Typography, Button, useMediaQuery } from '@mui/material';
+import { Typography, Button } from '@mui/material';
 import { useSelector, useDispatch } from 'react-redux';
-import { useTheme } from '@emotion/react';
-import {
-  ResponsiveContainer,
-  BarChart,
-  XAxis,
-  Bar,
-  YAxis,
-  LabelList,
-} from 'recharts';
 import Accordion from '@mui/material/Accordion';
 import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import {
-  twoDigit, reviewMix, reviewMixNECCC, calculatePieChartData,
+  reviewMix, reviewMixNECCC, calculatePieChartData,
   calculatePlantsandSeedsPerAcre,
 } from '../../../../shared/utils/calculator';
 import ReviewMixSteps from './Steps';
-import { DSTPieChart } from '../../../../components/DSTPieChart';
+import DSTPieChart from '../../../../components/DSTPieChart';
+import DSTBarChart from '../../../../components/DSTBarChart';
 import SeedingRateCard, { UnitSelection } from '../../../../components/SeedingRateCard';
 import { setBulkSeedingRateRedux, setOptionRedux } from '../../../../features/calculatorSlice/actions';
 import '../steps.scss';
@@ -58,11 +50,7 @@ const defaultPieChartData = {
 
 // eslint-disable-next-line no-unused-vars
 const ReviewMix = ({ calculator }) => {
-  const theme = useTheme();
-  const matchesSm = useMediaQuery(theme.breakpoints.down('sm'));
-
   const dispatch = useDispatch();
-
   const { council } = useSelector((state) => state.siteCondition);
   const {
     seedsSelected, sideBarSelection, options,
@@ -177,96 +165,6 @@ const ReviewMix = ({ calculator }) => {
   //                     Render                           //
   /// ///////////////////////////////////////////////////////
 
-  const renderAccordianChart = (seed) => {
-    const labels = [
-      {
-        label: 'Single Species Seeding Rate',
-        caption: 'Initial seeding rate',
-        val: calculatorResult[seed.label].step1.singleSpeciesSeedingRate,
-      },
-      {
-        label: 'Added to Mix',
-        caption: 'Mix proportion',
-        val: calculatorResult[seed.label].step2.seedingRate,
-      },
-      {
-        label: 'Drilled or Broadcast with Cultipack',
-        caption: 'Seeding method',
-        val: calculatorResult[seed.label].step2.seedingRateAfterPlantingMethodModifier,
-      },
-      {
-        label: `Management Impacts on Mix (${calculatorResult[seed.label].step3.managementImpactOnMix})`,
-        caption: 'Management',
-        val: calculatorResult[seed.label].step3.seedingRateAfterManagementImpact,
-      },
-      {
-        label: 'Bulk Germination and Purity',
-        caption: 'Bulk Germination and Purity',
-        val: calculatorResult[seed.label].step4.bulkSeedingRate,
-      },
-    ];
-
-    // eslint-disable-next-line react/no-unstable-nested-components
-    const CustomTick = ({ x, y, payload }) => (
-      <g transform={`translate(${x},${y})`}>
-        <text
-          textAnchor="middle"
-          fill="#666"
-          x={0}
-          y={10}
-          style={{ fill: '#4f5f30', whiteSpace: 'normal' }}
-        >
-          {`${payload.value}`}
-        </text>
-      </g>
-    );
-
-    return (
-      <Grid container>
-        {!matchesSm
-        && (
-        <Grid item xs={12}>
-          <ResponsiveContainer width="100%" height={200}>
-            <BarChart
-              data={labels}
-              barCategoryGap="25%"
-              margin={{
-                top: 15, right: 30, bottom: 15, left: 15,
-              }}
-            >
-              <XAxis
-                dataKey="caption"
-                interval={0}
-                tick={<CustomTick />}
-              />
-              <YAxis />
-              <Bar dataKey="val" fill="#4f5f30">
-                <LabelList dataKey="val" position="top" color="white" style={{ fill: '#4f5f30' }} />
-              </Bar>
-            </BarChart>
-          </ResponsiveContainer>
-        </Grid>
-        )}
-
-        {labels.map((l, i) => (
-          <Grid
-            container
-            sx={{ backgroundColor: !(i % 2) && '#e3e5d3' }}
-            key={i}
-          >
-            <Grid item xs={1} />
-            <Grid item sx={{ textAlign: 'justify' }} xs={9} pl={1}>
-              {l.caption}
-            </Grid>
-            <Grid item xs={2}>
-              {twoDigit(l.val)}
-            </Grid>
-          </Grid>
-        ))}
-      </Grid>
-    );
-  };
-
   return (
     <Grid container>
       <Grid item xs={12}>
@@ -312,10 +210,10 @@ const ReviewMix = ({ calculator }) => {
               <Typography>{seed.label}</Typography>
             </AccordionSummary>
             <AccordionDetails>
-              {renderAccordianChart(seed)}
 
-              <Grid container pt="1rem">
-                <Grid item xs={6}>
+              <Grid container>
+                <DSTBarChart seed={seed} calculatorResult={calculatorResult} />
+                <Grid item xs={6} pt="1rem">
                   <SeedingRateCard
                     seedingRateLabel="Seeding Rate in Mix PLS"
                     seedingRateValue={calculatorResult[seed.label].step2.seedingRate}
@@ -324,7 +222,7 @@ const ReviewMix = ({ calculator }) => {
                   />
                 </Grid>
 
-                <Grid item xs={6}>
+                <Grid item xs={6} pt="1rem">
                   <SeedingRateCard
                     seedingRateLabel="Bulk Seeding Rate"
                     seedingRateValue={calculatorResult[seed.label].step4.bulkSeedingRate}
